+++ .github/workflows/npm-publish.yml
@@ -1,0 +1,50 @@
+name: NPM Publish
+
+on:
+  push:
+    tags:
+      - '*'
+  workflow_dispatch:
+    inputs:
+      tag-name:
+        description: Tag to create release from
+        required: true
+        type: string
+        default: 2.1.0
+
+env:
+  NODE_VERSION: 18.x
+
+jobs:
+  npm-publish:
+    name: NPM Publish
+    runs-on: ubuntu-latest
+    steps:
+      - name: Get tag
+        if: ${{ github.event_name != 'workflow_dispatch' }}
+        run: echo "TAG_NAME=$(echo $GITHUB_REF | cut -d / -f 3)" >> $GITHUB_ENV
+
+      - name: Checkout repository
+        uses: actions/checkout@v3
+
+      - uses: actions/setup-node@v4
+        with:
+          node-version: ${{ env.NODE_VERSION }}
+          registry-url: https://registry.npmjs.org/
+
+      - name: Install dependencies
+        run: npm install
+
+      - name: Build package
+        run: npm run build
+
+      - name: Update npm version
+        run: npm version $VERSION --no-git-tag-version --allow-same-version
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_QA_TOKEN }}
+          VERSION: ${{ env.TAG_NAME }}
+
+      - name: Publish
+        run: npm publish
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_QA_TOKEN }}
