+++ e2e/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "ws_client_e2e_tests",
+  "version": "0.0.0",
+  "private": true,
+  "scripts": {
+    "test:e2e": "jest --config=jest.config.json",
+    "test:e2e:watch": "jest --config=jest.config.json --watchAll",
+    "test:e2e:detectOpenHandles": "jest --config=jest.config.json --detectOpenHandles"
+  },
+  "devDependencies": {
+    "@jest/globals": "^29.7.0",
+    "@types/jest": "^29.5.10",
+    "@types/node": "^20.10.0",
+    "@types/supertest": "^2.0.16",
+    "jest-expect-message": "^1.1.3",
+    "jest-extended": "^4.0.2",
+    "jest-mock-extended": "^3.0.5",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.3.2",
+    "ramda": "^0.29.1",
+    "@types/ramda": "^0.29.1"
+  },
+  "engines": {
+    "yarn": "please use pnpm",
+    "npm": "please use pnpm",
+    "node": ">=18",
+    "pnpm": ">=8"
+  }
+}
